Add Daniel Mack's SquashFS performance improvement
This patch dramatically speeds up SquashFS performance. This has a few
caveats:

1. You need to re-flash the MicroSD card/eMMC to get it since it's a
   U-Boot change.
2. The current fwup.conf for the BBB avoids reading large files (like
   the Linux kernel from SquashFS) because it was slow.

New custom systems based on the BBB can benefit from the simplicity of
storing the Linux kernel in the root filesystem and removing the need
for a FAT filesystem. This is, unfortunately, a backward incompatible
change for nerves_system_bbb so it can't be done here.