Enable /proc/config.gz
This makes it possible to inspect the configuration used to
compile the Linux kernel. This makes it possible for libraries that rely
on Linux kernel options being enabled to be able to verify that they are
enabled.

One use case is for `VintageNet`. `VintageNet` relies on many options
being enabled. Some of these options are subtle. For example,

1. Systems without `CONFIG_MULTICAST=y` sometimes can advertise their
   mDNS name and sometimes can't.
2. Route table configuration requires `CONFIG_IP_MULTIPLE_TABLES=y`
3. `vintage_net_mobile` requires `CONFIG_PPP=m`
4. VintageNet's Quectel modem support requires
   `CONFIG_USB_NET_QMI_WWAN=m`

While this knowledge is documented, mostly, I think that it's pretty
hard to find since the errors don't lead you to it. I don't even know
how to do that without a lot of work. The easier route seems to be to
have a function that reviews the `config.gz` and prints out warnings and
errors. This change enables that function.

This adds about 20 KB to the kernel. This is about 1% of the size of the
Raspberry Pi Zero's Linux kernel.